# spec/openapi.yaml
openapi: 3.1.0
info:
  title: PAL NAS Media Uploader API
  version: 1.0.0
  description: |
    Backend API for a self-hosted, multi-user NAS media library.

    ## Accounts and authorization

    The service has two roles. Administrators manage accounts but cannot use
    upload or gallery routes. Regular users upload into their configured folder,
    organize accessible media, and share owned media with other regular users.
    Resource lookup is scoped by the authenticated identity; inaccessible
    resources generally return `404` instead of revealing their existence.

    The first persistent administrator is created through the one-time HTML
    setup flow. Deployment bootstrap credentials authorize that operation only
    and are never stored as an application account. The initial administrator
    enrolls TOTP on first login. Every account subsequently created by an
    administrator also uses password plus TOTP.

    ## Sessions and browser access

    Authentication produces an opaque, HttpOnly cookie named
    `pal_medias_uploader_session`. Sessions last 30 days. Cross-origin browser
    requests must send credentials and their `Origin` must appear in
    `ENV_CORS_ALLOWED_ORIGINS`; an unlisted origin receives an empty `403`
    response. The cookie is `Secure` for HTTPS requests and uses `SameSite=None`
    only for secure cross-site requests; otherwise it uses `SameSite=Strict`.

    ## Uploads and media

...

...

    ## Albums

...
    have membership changed manually.

...

...

...

...
servers:
...
security:
  - sessionCookie: []
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      operationId: beginLogin
      summary: Log in with a password and optional TOTP
      description: |
        Accepts a JSON body of at most 1 MiB and rejects unknown fields. The
        username is trimmed and matched case-sensitively; the password is checked
        with bcrypt.

        The initial configured administrator receives a session cookie
        immediately and the response contains `authenticated: true`. Every
        subsequently created account receives an opaque challenge valid for five
        minutes. Managed accounts receive their TOTP provisioning URI when an
        administrator creates them. The login response exposes provisioning data
        only if an account has a pending, not-yet-active TOTP secret.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: A TOTP challenge was created, or the initial administrator was authenticated immediately.
          headers:
            Set-Cookie:
              description: Set immediately for accounts exempt from TOTP.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginChallenge"
              examples:
                initialAdministrator:
                  summary: Password-only initial administrator
                  value:
                    authenticated: true
                totpChallenge:
                  summary: Existing TOTP-enabled account
                  value:
                    challengeToken: opaque-challenge-token
                enrollmentChallenge:
                  summary: First login for a managed account
                  value:
                    challengeToken: opaque-challenge-token
                    provisioningUri: otpauth://totp/PAL:user
                    secret: JBSWY3DPEHPK3PXP
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenOrigin"
        "409":
          $ref: "#/components/responses/UploadConflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /auth/verify:
    post:
      tags: [Authentication]
      operationId: verifyTOTP
      summary: Verify TOTP and create a session
      description: |
        Validates an unexpired challenge and a six-digit TOTP code. TOTP uses a
        30-second period, SHA-1, six digits, and accepts one adjacent time step in
        either direction. For first-time login, successful verification promotes
        the pending secret to the account's active secret. The challenge is then
        deleted and cannot be reused.

        A successful response creates a server-side session valid for 30 days
        and sets the opaque HttpOnly session cookie.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyTOTPRequest"
      responses:
        "200":
          description: Authenticated session created.
          headers:
            Set-Cookie:
              description: HttpOnly session cookie named pal_medias_uploader_session.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticatedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenOrigin"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /auth/recovery:
    post:
      tags: [Authentication]
      operationId: verifyRecoveryCode
      summary: Authenticate with a single-use TOTP recovery code
      description: Requires the short-lived challenge returned after successful password verification. Three invalid attempts are rate limited with the same controls as TOTP verification.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoveryVerificationRequest"
      responses:
        "200":
          description: Recovery code consumed and authenticated session created.
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticatedResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          description: Too many authentication attempts.
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pal_medias_uploader_session
      description: |
        Opaque session token stored only as a hash by the server. It is issued
        after successful TOTP verification, or immediately for the initial
        administrator, and expires after 30 days. Browser cookies are HttpOnly,
        use path `/`, and are Secure whenever the request is HTTPS.
...
  schemas:
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          description: Case-sensitive username. Surrounding whitespace is ignored.
          examples: [admin]
        password:
          type: string
          format: password
          minLength: 1
          description: Plaintext password transported over HTTPS and checked with bcrypt.
    LoginChallenge:
      type: object
      additionalProperties: false
      description: Exactly one shape is returned, immediate authentication or a TOTP challenge.
      oneOf:
        - required: [authenticated]
          properties:
            authenticated:
              const: true
        - required: [challengeToken]
      properties:
        challengeToken:
          type: string
          minLength: 1
          description: Opaque challenge valid for five minutes.
        provisioningUri:
          type: string
          format: uri
          description: Present only while a managed account requires initial TOTP enrollment.
        secret:
          type: string
          pattern: "^[A-Z2-7]+=*$"
          description: Base32 TOTP secret present only during initial enrollment. Store it securely.
        authenticated:
          type: boolean
          description: True only when password login immediately created a session for the initial administrator.
    VerifyTOTPRequest:
      type: object
      additionalProperties: false
      required: [challengeToken, code]
      properties:
        challengeToken:
          type: string
          minLength: 1
        code:
          type: string
          pattern: "^[0-9]{6}$"
    AuthenticatedResponse:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          const: authenticated
...
    RecoveryVerificationRequest:
      type: object
      additionalProperties: false
      required: [challengeToken, code]
      properties:
        challengeToken:
          type: string
          minLength: 1
        code:
          type: string
          pattern: "^[A-Za-z2-7 -]+$"
          description: One of the generated Base32 recovery codes. Hyphens, spaces, and letter case are ignored.
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
          description: Stable or human-readable error message. Clients should primarily branch on HTTP status.
  responses:
    BadRequest:
      description: Malformed JSON or unknown authentication request field.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            malformed:
              value:
                error: Bad Request
...
    Unauthorized:
      description: Missing or invalid session, credentials, challenge, or TOTP code.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            unauthorized:
              value:
                error: Unauthorized
    ForbiddenOrigin:
      description: The Origin header is not in the comma-separated `ENV_CORS_ALLOWED_ORIGINS` allowlist. Response has no body.
...
    UploadConflict:
      description: The checksum already belongs to owned media, or at least one expected upload part is missing.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            duplicate:
              value:
                error: media with this checksum already exists
            missingPart:
              value:
                error: missing part 2
    InternalServerError:
      description: Unexpected filesystem, database, token generation, or encoding failure.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            internal:
              value:
                error: internal server error